requirements.txt
Django
djangorestframework
drf-spectacular
django-filter
djangorestframework-simplejwt
requests
